# docs/openapi.yaml
openapi: 3.0.3
info:
  title: NeuroClinAIcal Backend API
  version: 1.0.0
  description: API for managing user accounts, patients, and medical reports for NeuroClinAIcal.

# servers:
#   - url: TODO
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
  /user/login:
    post:
      summary: Log in a user
      description: Authenticates a user and returns a JWT token for accessing protected endpoints.
      operationId: login_user
      tags:
        - User
      security: [] # Public
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  description: The email address of the user.
                  example: "johndoe@example.com"
                password:
                  type: string
                  format: password
                  description: The password of the user.
                  example: "password123"
      responses:
        "200":
          description: Successful login. Returns a JWT token.
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: The JWT token for authentication.
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        "401":
          description: Invalid credentials.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Invalid email or password."
  /user/register:
    post:
      summary: Register a new user
      description: Creates a new user account in the system.
      operationId: register_user
      tags:
        - User
      security: [] # Public
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - name
                - email
                - password
              properties:
                username:
                  type: string
                  example: "johndoe123"
                name:
                  type: string
                  example: "John Doe"
                email:
                  type: string
                  format: email
                  example: "johndoe@example.com"
                password:
                  type: string
                  format: password
                  example: "password123"
      responses:
        "201":
          description: User created successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 123
                  username:
                    type: string
                    example: "johndoe123"
                  name:
                    type: string
                    example: "John Doe"
                  email:
                    type: string
                    example: "johndoe@example.com"
                  token:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        "400":
          description: Bad request.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Email is already in use."
...